refactor: move mesher to own file
